Adding .nim versions
Adding .nim versions of my written tutorial versions for easier linking of project issues. Also, it's pointless not to, since nimib by default has a show source at the bottom.

Also removing rename of WIP 1, 2, 3 to their original names, we are now tracking which one to do in order using the project dashboard.