Fix and Update for the SDL2 Video Series
Without this, the viewer must look at the comments for the solution of the problem of differently named import due to a newer version of sdl2_nim.

ALSO, every single code link to the playground has been broken for years now. Someone could of told me that, how am i the only one to notice ?...